requirements.txt
Django==5.0.4
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
python-decouple==3.8


# psycopg2-binary==2.9.9 --only-binary=:all:

